+++ CHANGELOG.md
@@ -8,6 +8,7 @@ master
 * Added menu item to search for updates. Fix #54.
 * Log with Growl on 10.7, with notification center on 10.8 and up. Fix #59.
 * Use windowDidLoad rather than override showWindow. Fix #62.
+* Never die while processing a screenshot, log exceptions as needed. Fix #38.
 
 v2.0 (8/22/2014)
 ----------------